fix: fixed pull request review suggestions and exercise pop up styles
