+++ src/css/common.css
@@ -1,6 +1,9 @@
 :root {
   /* Colors */
   --light: #f4f4f4;
+  --light-20: #f4f4f433;
+  --light-40: #f4f4f466;
+  --light-60: #f4f4f499;
   --light-grey: #d4d4d4;
   --light-blue-grey: #c6cdd1;
   --dark: #242424;
