+++ src/partials/exercise_pop_up.html
@@ -1,4 +1,4 @@
-<div class="exercise-pop-up-backdrop js-exercise-pop-up-container is-open">
+<div class="exercise-pop-up-backdrop js-exercise-pop-up-container">
   <div class="exercise-pop-up-container" id="exercise-pop-up">
     <button class="exercise-pop-up-close-btn js-close-exercise-pop-up" aria-label="Close exercise">
       <svg class="exercise-pop-up-close-icon" width="24" height="24">
